requirements.txt
google-generativeai>=0.3.0
fpdf2>=2.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0 